send a last PDU when closing a not fully consumed Stream
If a handler returns without consuming and/or closing the reqStream,
we want the client to know that the server had to close because of the
handler rather than an EOF which looks exactly like a crash.

For Client, though, we return *Stream into the wild, and have to rely on
the Client user's discipline to close the returned *Stream before
calling another RequestReply on Client.
We enforce that through Conn.recvBusy.