make timeouts simple inactivity timeouts
The problem with the previous implementation was that the client would
start sending, but also receive simultaneously.
If the send took longer than RxTimeout, the concurrent recv would return
with an error, and the client would close the connection, even though
everything was perfectly fine (send was not finished, so no response to
be expected yet).

By using SetDeadline, we avoid that problem, but the timeout now is a
simple inactivity timeout, which is reflected in streamrpc.ConnConfig.